Added linux branch

+++ Train_Eval/tools/datasets/relabel_embryo_phenotype.py
@@ -20,7 +20,7 @@ def main():
     parser.add_argument('--start',"-st", type=int, default=640)
     parser.add_argument('--severe',"-se", type=int, default=100)
     parser.add_argument('--class_name',"-cn", type=str, default="NORMAL")
-    parser.add_argument('--ignore_class', "-ic",  nargs="+", default=["NORMAL", "BOOM", "CUT"])
+    parser.add_argument('--ignore_class', "-ic",  nargs="+", default=["BOOM", "CUT"])
     parser.add_argument('--timestamp_prefix', "-tp", type = str, default= "LO")
     args = parser.parse_args()
     class_name = args.class_name.upper()
@@ -37,39 +37,42 @@ def main():
         path_to_experiment = os.path.join(path_to_session, experiment)
         path_to_json_folder = os.path.join(path_to_experiment, labeler_name)
 
-        path_to_save_annotation = os.path.join(path_to_experiment, updated_annotation_name)
-        os.makedirs(path_to_save_annotation, exist_ok=False)
-
-        for json_file in os.listdir(path_to_json_folder):
-            age = re.findall(timestamp_prefix+ "\d+", json_file)
-            age = int(age[0][len(timestamp_prefix):])
-            path2Json = os.path.join(path_to_json_folder, json_file)
-            path_to_save_annotation_file = os.path.join(path_to_save_annotation, json_file)
-            with open(path2Json,) as f:
-                annotation = json.load(f)
-
-                for detection in annotation["detection_list"]:
-                    if args.start == 0:
-                        if detection not in ignore_class_names:
-                            detection["className"] = class_name
-                            detection["severe"] = args.severe
-
-
-                    if (args.start > 0) and (args.start < age):
-                        if detection["className"] not in ignore_class_names:
-                            detection["className"] = class_name
-                            detection["severe"] = args.severe
-                        if detection["className"] in (ignore_class_names + ["UNKNOWN"]):
+        if os.path.isdir(path_to_experiment) and os.path.exists(path_to_json_folder):
+    
+            path_to_save_annotation = os.path.join(path_to_experiment, updated_annotation_name)
+            if os.path.exists(path_to_save_annotation) == False:
+                os.makedirs(path_to_save_annotation, exist_ok=False)
+    
+            for json_file in os.listdir(path_to_json_folder):
+                age = re.findall(timestamp_prefix+ "\d+", json_file)
+                age = int(age[0][len(timestamp_prefix):])
+                path2Json = os.path.join(path_to_json_folder, json_file)
+                path_to_save_annotation_file = os.path.join(path_to_save_annotation, json_file)
+                with open(path2Json,) as f:
+                    annotation = json.load(f)
+    
+                    for detection in annotation["detection_list"]:
+                        if args.start == 0:
+                            if detection not in ignore_class_names:
+                                detection["className"] = class_name
+                                detection["severe"] = args.severe
+    
+    
+                        if (args.start > 0) and (args.start < age):
+                            if detection["className"] not in ignore_class_names:
+                                detection["className"] = class_name
+                                detection["severe"] = args.severe
+                            if detection["className"] in (ignore_class_names + ["UNKNOWN"]):
+                                detection["severe"] = 100
+                        else:
                             detection["severe"] = 100
-                    else:
-                        detection["severe"] = 100
-
-
-                updated_annotation = annotation
-
-
-            with open(path_to_save_annotation_file,"w") as fanno:
-                json.dump(updated_annotation, fanno, indent=4)
+    
+    
+                    updated_annotation = annotation
+    
+    
+                with open(path_to_save_annotation_file,"w") as fanno:
+                    json.dump(updated_annotation, fanno, indent=4)
 
 
 if __name__=="__main__":
